backend/requirements.txt
Django==4.0.7
django-annoying
django-cors-headers
django-extensions
django-polymorphic
djangorestframework
djangorestframework
djangorestframework_simplejwt
djoser
matplotlib
psycopg2-binary
uwsgi
Werkzeug
redis
hiredis
names
